fix: address Oracle review findings for Pexels image picker
Critical:
- Preserve MaterialInfo.material_type when the picker reconstructs
  selected_materials; without it every image picked in the UI was
  routed through save_video in the pipeline and broke every
  image-selection task.

Important:
- Use float clip_durations consistently between the download and
  combine phases so per-paragraph audio/video stay in sync; widen
  render_image_zoom_video and download_selected_videos to accept
  list[float] (was list[int]).
- Guard combine-phase parse_subtitle_cues against missing subtitle
  files (FileNotFoundError/OSError) with a logged fallback, matching
  the silent tolerance in the download phase.

Minor:
- Add PIL.Image.DecompressionBombError to the catch tuple in
  _download_image_file so decompression-bomb payloads surface as
  the helper's documented 'returns None' contract instead of
  escaping to the outer broad except.
- Sanitize URLs (_safe_public_url) in the new log lines added by
  the image download helper.
- Add zh.json translations for the 9 new i18n keys.

Test:
- Add material_type attribute to _FakePexelsOption in
  test_webui_paragraph_picker so the existing video test continues
  to drive _render_generation_controls end-to-end.

+++ app/services/video.py
@@ -1388,7 +1388,7 @@ def generate_video(
         return bgm_mix_succeeded
 
 
-def render_image_zoom_video(image_path: str, clip_duration: int = 5) -> str:
+def render_image_zoom_video(image_path: str, clip_duration: float = 5) -> str:
     """
     将单张本地图片渲染为带缓慢放大效果的 mp4 片段，返回输出文件路径。
